Rename: GridList -> GridType
